+++ pkg/ui/prompts.go
@@ -22,3 +22,20 @@ func YesNoPrompt(message string, defaultValue bool) bool {
 
 	return answer
 }
+
+func MultiSelectPrompt(message string, options, defaults []string) []string {
+	if viper.GetBool(ASSUME_YES_FLAG) {
+		return defaults
+	}
+
+	prompt := &survey.MultiSelect{
+		Options: options,
+		Default: defaults,
+		Message: message,
+	}
+
+	var response []string
+	survey.AskOne(prompt, &response)
+
+	return response
+}